refactor: 변수, 함수 네이밍 수정 및 파일 분리
